+++ dio_test/lib/src/test/status_code_tests.dart
@@ -13,7 +13,7 @@ void statusCodeTests(
 
   group('status code', () {
     for (final code in [400, 401, 404, 500, 503]) {
-      test('$code', () async {
+      test('$code', () {
         expect(
           dio.get('/status/$code'),
           throwsDioException(
@@ -28,5 +28,34 @@ void statusCodeTests(
         );
       });
     }
+  });
+
+  group(ValidateStatus, () {
+    test('200 with validateStatus => false', () {
+      expect(
+        dio.get(
+          '/status/200',
+          options: Options(validateStatus: (status) => false),
+        ),
+        throwsDioException(
+          DioExceptionType.badResponse,
+          stackTraceContains: kIsWeb ? null : 'test/status_code_tests.dart',
+          matcher: isA<DioException>().having(
+            (e) => e.response!.statusCode,
+            'statusCode',
+            equals(200),
+          ),
+        ),
+      );
+    });
+
+    test('500 with validateStatus => true', () async {
+      final response = await dio.get(
+        '/status/500',
+        options: Options(validateStatus: (status) => true),
+      );
+
+      expect(response.statusCode, 500);
+    });
   });
 }